+++ src/main/java/com/telepathicgrunt/the_bumblezone/entities/BeeInteractivity.java
@@ -20,7 +20,6 @@ import net.minecraft.util.SoundCategory;
 import net.minecraft.util.SoundEvents;
 import net.minecraft.world.World;
 import net.minecraft.world.server.ServerWorld;
-import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.registries.ForgeRegistries;
 
 public class BeeInteractivity {